Correctly set the command interpreter for sending commands to bash, rather than /bin/sh.
`/bin/sh` is the bourne shell. `bash` is the bourne again
shell. `sh` does not actually support the `set -uexo pipefail`
command, as `-o pipefail` is a bashism, not found in the
bourne shell or any shell that is strictly compatible with it.

This allows the build to be successfully executed on
Ubuntu, and other distros that symlink /bin/sh to something
other than bash.